backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
groq==0.11.0
httpx==0.27.0
requests==2.31.0
reportlab==4.0.7
Pillow==10.1.0
python-dotenv==1.0.0
transformers>=4.45.0
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.25.0
qwen-vl-utils
sentencepiece
diffusers>=0.25.0
safetensors>=0.4.0
omegaconf>=2.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
email-validator==2.1.0
